### 2022-06-03 #4450 @editorialbot
[PRE REVIEW]: UnROOT: an I/O library for the CERN ROOT file format written in Julia
**Submitting author:** @tamasgal (<a href="http://orcid.org/0000-0001-7821-8673">Tamás Gál</a>)
**Repository:** https://github.com/JuliaHEP/UnROOT.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.8.8
