+++ paper/paper.bib
@@ -17,6 +17,7 @@
   journal={Computing and Software for Big Science},
   volume={5},
   number={1},
+  doi = {10.1007/s41781-021-00053-3},
   pages={1--11},
   year={2021},
   publisher={Springer}
