### 2022-06-15 #4452 @danielskatz
Hi @PerilousApricot and @jpata - I just wanted to check in and see how things are going...
