### 2022-06-22 JuliaHEP/UnROOT.jl#170 @jpata
[JOSS review] Contribution guidelines
As part of the ongoing [JOSS review](https://github.com/openjournals/joss-reviews/issues/4452), for the item:

  - **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

it might be worth to add a few more details in the README / docs. For example:
- should the "issues" tab be the main place for discussion and support, or do you have another channel in mind?
- if someone has code changes in mind, should they just open a PR, or should it be discussed first at some venue?
