### 2022-06-22 JuliaHEP/UnROOT.jl#171 @jpata
[JOSS review] Paper comments
Congrats on the substantial work on the package and the paper! I had some comments in the context of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4452) which would be good to address before I sign off.
I'm still working through some actual practical use cases in code, so I might have some more comments later.

## Content-related comments:

1. I understand the goal is to provide performant access to ROOT files in a non-vectorized (for-loop) format, in Julia. It might be useful to add some concrete benchmarks to the paper to compare this package (Julia for-loop) to the alternatives:
    - a C++ ROOT loop
    - a C++ RDataFrame vectorized analysis
    - a python/awkward-array vectorized analysis
2. the interface supports multi-threading which is great, do you have any information on the multi-threading performance (e.g. scaling) and correctness?
3. lines 21-23: "dancing across language barriers hinders the ability to parallelize tasks that are conceptually trivial most of the time". I'm not sure if this is universally true. Running parallel code from python via e.g. multiprocessing or numba is often quite simple and effective. I think also parallelized RDataFrame via Python should be fairly easy to use and performant.
4. line 49 states that UnROOT is used in CMS analysis, citing https://doi.org/10.1051/epjconf/202024506002. Checking those proceedings, there is no mention of Julia or UnROOT, therefore, I'm not sure if this is a statement supported by this citation. I checked the other citation for KM3NeT, and there is also no mention of julia/UnROOT there (it's from 2016, which predates uproot/UnROOT I believe).
5. lines 56-57 mention composability with other nice Julia packages like queries or loop fusion, it would be great to put some example code in the paper!
6. line 79 in the Summary states that the processing speeds from UnROOT are comparable to the C++ root framework, but there are no concrete benchmark results in the paper to support this claim (see also point 1)

## Style-related comments:
- line 11: "HEP community as been troubled by the two-language problem for a long time" -> This statement is rather generic. Who decides that this is actually a problem? Should one language solve everything, from configuration to computation on different devices and platforms?
- line 14:  "vectorized style, a type of problems which are normally tackled with" -> "usually implemented with ...". I think already here awkward-array (with dedicated precompiled kernels) should be mentioned and cited as a concrete solution to the problem of doing vectorized computations on jagged data.
- line 18: "physicists who usually have no or little background in software engineering" -> is this statement suppported by any studies? In any case, I would imagine that the physicists who are interested to try Julia (as opposed to using something used by more people) might need to be a bit more software-inclined than the average.
- line 40: resemble -> represent
- ~~for awkward-array, in addition to the zenodo of the software, it might be worth it to cite some of these proceedings [10.1051/epjconf/201921406026](https://doi.org/10.1051/epjconf/201921406026), [10.1051/epjconf/202024505023](https://doi.org/10.1051/epjconf/202024505023), [10.1088/1742-6596/1525/1/012053](https://doi.org/10.1088/1742-6596/1525/1/012053)~~ (according to awkward authors, it's preferred to cite just the software)
