### 2022-06-22 #4452 @jpata
From my side, I'm done with the initial paper review and have left some issues:
- Contribution guidelines: https://github.com/JuliaHEP/UnROOT.jl/issues/170
- Adding some more details and improving the writing in the paper: https://github.com/JuliaHEP/UnROOT.jl/issues/171

I'm also going to try out the library in the next days and check off the functionality/performance based on that.
