### 2022-06-23 JuliaHEP/UnROOT.jl#171 @Moelf
thanks for the review, regarding `1.` do you think this is good enough:
https://github.com/Moelf/UnROOT_RDataFrame_MiniBenchmark
