### 2022-06-23 527f04b
Add `view` of LazyTree (#140)
