+++ test/runtests.jl
@@ -302,6 +302,24 @@ end
     close(rootfile)
 end
 
+@testset "View" begin
+    data = LazyTree(joinpath(SAMPLES_DIR, "tree_with_jagged_array.root"), "t1")
+    data[1:2]
+    @view data[1:2]
+    alloc1 = @allocated v = data[3:90]
+    alloc2 = @allocated v = @view data[3:90]
+    v = @view data[3:80]
+    @test alloc2 < alloc1/100
+    @static if VERSION >= v"1.8"
+        @test alloc2 < 50
+    end
+    @test all(v.int32_array .== data.int32_array[3:80])
+
+    v2 = @view data[[1,3,5]]
+    @test v2[1].int32_array == data[1].int32_array
+    @test v2[2].int32_array == data[3].int32_array
+end
+
 @testset "Doubly jagged branches" begin
     rootfile = ROOTFile(joinpath(SAMPLES_DIR, "tree_with_doubly_jagged.root"))
     vvi = [[[2], [3, 5]], [[7, 9, 11], [13]], [[17], [19], []], [], [[]]]
