### 2022-06-26 JuliaHEP/UnROOT.jl#171 @tamasgal
Thanks for your valuable feedback and thanks @Moelf for the quick fixes in the other PR `:)`

Regarding 4) we cited the experiments themselves and not direct publications which involve `UnROOT`, maybe thatwas a bit unclear or misreading. `UnROOT` is used in many analyses in KM3NeT (as a core library) but I think it's not explicitly mentioned (yet). One of the goals of the JOSS paper submission was to get a DOI and encourage people to mention `UnROOT` after all ;)
For CMS I think it's a fairly similar situation, but @Moelf will comment for sure.
Should we make it more clear or leave it out?

5) Yes, we should include examples, that's a good idea indeed.

1) @Moelf I think your microbenchmarks are well suited for this. Those were also discussed with other experts so I think it's fine! Let's see what @jpata thinks about it.

Thanks for the style related comments, yes there is room for improvement and clarity. I will work on that...
Regarding line 18: we don't have any citation but it's a fairly common observation among many different institutes in my experience. I agree, it suggests it's a backed statement, I'll rephrase that too `;)`
