### 2022-06-30 JuliaHEP/UnROOT.jl#171 @jpata
@Moelf
> thanks for the review! do you think this is good enough: https://github.com/Moelf/UnROOT_RDataFrame_MiniBenchmark

I think quoting these numbers (potentially with a link to the repo) gives a good impression that it's a quantitative statement.

> The scaling from 1's benchmark shows linear, just like g++ compiled RDataFrame

That's great - I did some tests too and after 2-4 threads I didn't see too much improvement, but probably depends on the circumstances. Have you checked that the code running with several threads returns exactly the same results (i.e. there is no multithreading bugs or dependence on non-associative floating point operations)?

> it seems python is 2x slower. (I personally suspect it's due to JIT optimization being lower for interactive use)

Ha, good to know...

In any case, from my point of view, the performance points 1-2 are addressed.

@tamasgal
> For CMS I think it's a fairly similar situation, but @Moelf will comment for sure. Should we make it more clear or leave it out?

If you want to just give a reference to CMS, citing NanoAOD proceedings is probably a bit of an odd choice. Perhaps a more truthy statement would be that UnROOT, like RDataFrame, can directly be used with flat ntuples such as CMS NanoAOD? In any case, it'd be nice to hear if it's really being used in an analysis (even if a citation is not available for understandable reasons)!
For KM3NeT, the citation and the statement are also confusing, so it might be good to reword to make it clearer.

Note that I will be mostly off in July, but once the other comments are addressed, I can sign off from my side.
