+++ paper/paper.md
@@ -30,8 +30,9 @@ bibliography: paper.bib
 (`.root`) that is fast, memory-efficient, and composes well with Julia's
 high-performance iteration, array, and multi-threading interfaces.
 
-# Statement of need
-The High-Energy Physics (HEP) community has been troubled by the two-language
+# Statement of Need
+The High-Energy Physics (HEP) community, especially the group doing analysis, 
+has facing the two-language
 problem for a long time. Often, physicists would start prototyping with a
 `Python` front-end which glues to a `C/C++/Fortran` back-end. Soon they will hit
 a task which is extremely hard to express in columnar (i.e. "vectorized") style,
@@ -39,13 +40,21 @@ a type of problems which are normally tackled with libraries like
 `numpy` [@harris2020array] or `pandas` [@reback2020pandas]. This usually leads to
 either writing `C++` kernels and interface them with `Python`, or porting the
 prototype to `C++` and start to maintain two code bases including the wrapper
-code. Both options are engineering challenges for physicists who usually have no
-or little background in software engineering. Many steps of this process are
+code. Specific to HEP, `AwkwardArray` [@pivarski_jim_2018_6522027] can be seen
+as a compromise between the two solutions, where the user writes in a special
+columnar style that has some flexibility of addressing the jaggedness of HEP data.
+
+All traditional options represent incresing engineering effort for authors and 
+users, often in multiple programming languages. Many steps of this process are
 critical, like identifying bottlenecks, creating an architecture which is both
 performant and maintainable at the same time while still being user-friendly and
 logically structured. Using a `Python` front-end and dancing across language
-barriers also hinders the ability to parallelize tasks that are conceptually
-trivial most of the time.
+barriers also hinders the ability to parallelize tasks down to 
+event level, the existing usage often rely on chunk or even file level
+parallelization. Finally, newer techniques such as automatic differentiation
+also works more smoothly without language barriers, allows physicists to DIY
+algorithms themselves. With Julia's active auto diff community [^1], we expect 
+`UnROOT.jl` to be one of the pillar stone for physicists.
 
 `UnROOT.jl` attempts to solve all of the above by choosing Julia, a
 high-performance language with simple and expressive syntax [@Julia]. Julia is
@@ -69,7 +78,7 @@ there are some basic structures and constants needed to bootstrap the parsing
 process. One of the biggest advantages of the `ROOT` data format is the ability
 to store jagged structures like nested arrays of structs with different
 sub-array lengths. In high-energy physics, such structures are preferred to
-resemble e.g. particle interactions and detector responses as signals from 
+represent e.g. particle interactions and detector responses as signals from 
 different hardware components, combined into a tree of events.
 
 `UnROOT.jl` understands the core structure of `ROOT` files, and is able to
@@ -79,7 +88,9 @@ branches are supported as well, including their nested variants. Additionally,
 `UnROOT.jl` provides a way to hook into the deserialisation process of custom
 types where the automatic parsing fails. By the time of writing, `UnROOT` is
[…]
