+++ Project.toml
@@ -1,7 +1,7 @@
 name = "UnROOT"
 uuid = "3cd96dde-e98d-4713-81e9-a4a1b0235ce9"
 authors = ["Tamas Gal", "Jerry Ling", "Johannes Schumann", "Nick Amin"]
-version = "0.8.8"
+version = "0.8.9"
 
 [deps]
 AbstractTrees = "1520ce14-60c1-5f80-bbc7-55ef81b5835c"
@@ -34,7 +34,7 @@ CodecZstd = "^0.6.0, ^0.7.0"
 HTTP = "^0.9.7"
 IterTools = "^1"
 LRUCache = "^1.3.0"
-LazyArrays = "^0.15, ^0.21, ^0.22, ^1"
+LazyArrays = "^0.21, ^0.22, ^1"
 LibDeflate = "^0.4.1"
 LorentzVectors = "^0.4.0"
 Memoization = "^0.1.10"
