### 2022-07-04 bf36a65
Create CODE_OF_CONDUCT.md
