### 2022-07-04 JuliaHEP/UnROOT.jl#171 @Moelf
>That's great - I did some tests too and after 2-4 threads I didn't see too much improvement, but probably depends on the circumstances.

yeah, it's possible after some threading we're limited by I/O, and this comes in two forms:
1. the raw bandwidth of your I/O device under random read
2. the latency for (if HDD) reader to seek to a different location
