### 2022-07-07 acdef9e
Update ArraysOfArrays version bound
