### 2022-07-07 c1ce0ea
Increase package version to v0.8.10
