+++ Project.toml
@@ -31,7 +31,7 @@ ArraysOfArrays = "^0.5.3, ^0.6"
 CodecLz4 = "^0.3.0, ^0.4.0"
 CodecXz = "^0.6.0, ^0.7.0"
 CodecZstd = "^0.6.0, ^0.7.0"
-HTTP = "^0.9.7"
+HTTP = "^0.9.7, 1"
 IterTools = "^1"
 LRUCache = "^1.3.0"
 LazyArrays = "^0.21, ^0.22, ^1"
