### 2022-07-07 4c7a91a
CompatHelper: bump compat for AbstractTrees to 0.4, (keep existing compat)
