### 2022-07-15 #4452 @PerilousApricot
Errrr, apparently the Github email connector isn't working?? Sorry about that -- I *tried* to say earlier that I was on vacation until the 13th... I've nearly completed the review now and should be able to sign off on the paper in the next hour
