### 2022-07-15 JuliaHEP/UnROOT.jl#177 @PerilousApricot
Installation instructions in paper
Hello, in the following review: openjournals/joss-reviews#4452  .. the following information in the checklist appears to not be fulfilled

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Perhaps to a Julia developer it's clear, but as an outsider, I don't see where this is listed.

Great paper otherwise!
