### 2022-07-15 b1c4274
add installation guide
close #177
