### 2022-07-15 #4452 @danielskatz
Ah, @jpata, I see that you didn't check off an item: "**Quality of writing**: Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?"

What needs to be done for you to check this off?
