### 2022-07-15 #4452 @Moelf
I think when he comes back from vacation and take a look at the PR that closed:
- https://github.com/JuliaHEP/UnROOT.jl/issues/171

they probably can decide if we have more items to do
