### 2022-07-21 16a758e
Remove KM3NeT reference
