### 2022-08-09 #4452 @jpata
Sorry for the delay on my part due to vacations.

The paper has been improved significantly and all the content-related comments have been addressed, but I think it could still benefit from some language-level improvements. I'm not a native speaker, but here are some suggestions which could improve readability.

- `11. The High-Energy Physics (HEP) community, especially the group doing analysis, has facing the two-language problem for a long time. -> ... especially in data analysis, has been facing ...`
- `14. is extremely hard to express in columnar -> can be challenging to express`
- `16. leads to either writing C++ kernels and interface them -> ... and interfacing them`
- `26. the existing usage often rely on -> often relies on`
-  `26. newer techniques such as automatic differentiation also works more smoothly without language barriers, allows physicists to DIY algorithms themselves. -> ... also work more smoothly without language barriers, allowing physicists to develop algorithms.`
- `68. can be done with 0 code addition -> can be done with no code addition`

If you want to go forward as-is, I will be happy to sign off, but I think a little bit of polishing of the language could make the paper even better.
