### 2022-08-09 5943ade
address language comments from @jpata (#178)
Co-authored-by: Tamas Gal <himself@tamasgal.com>

[skip ci]
