+++ paper/paper.md
@@ -26,73 +26,73 @@ date: 08 October 2021
 bibliography: paper.bib
 ---
 # Summary
-`UnROOT.jl` is a pure Julia implementation of CERN ROOT [@Brun:1997pa] files I/O
-(`.root`) that is fast, memory-efficient, and composes well with Julia's
+`UnROOT.jl` is a pure Julia implementation of CERN's ROOT [@Brun:1997pa] file I/O
+(`.root`) software, which is fast and memory-efficient, and composes well with Julia's
 high-performance iteration, array, and multi-threading interfaces.
 
 # Statement of Need
 The High-Energy Physics (HEP) community, especially in data analysis, 
 has been facing the two-language
 problem for a long time. Often, physicists would start prototyping with a
-`Python` front-end which glues to a `C/C++/Fortran` back-end. Soon they will hit
-a task which can be challenging to express in columnar (i.e. "vectorized") style,
-a type of problems which are normally tackled with libraries like
-`numpy` [@harris2020array] or `pandas` [@reback2020pandas]. This usually leads to
-either writing `C++` kernels and interfacing them with `Python`, or porting the
-prototype to `C++` and start to maintain two code bases including the wrapper
+`Python` front-end which glues to a `C/C++/Fortran` back-end. Soon they would hit
+a task that could be challenging to express in columnar (i.e., "vectorized") style,
+a type of problem that is normally tackled with libraries like
+`numpy` [@harris2020array] or `pandas` [@reback2020pandas]. This usually would lead to
+them either writing `C++` kernels and interfacing them with `Python`, or porting the
+prototype to `C++` and starting to maintain two code bases including the wrapper
 code. Specific to HEP, `AwkwardArray` [@pivarski_jim_2018_6522027] can be seen
 as a compromise between the two solutions, where the user writes in a special
-columnar style that has some flexibility of addressing the jaggedness of HEP data.
+columnar style that has some flexibility for addressing the jaggedness of HEP data.
 
-All traditional options represent incresing engineering effort for authors and 
+All traditional options represent increasing engineering effort for authors and 
 users, often in multiple programming languages. Many steps of this process are
-critical, like identifying bottlenecks, creating an architecture which is both
-performant and maintainable at the same time while still being user-friendly and
+critical, such as identifying bottlenecks and creating an architecture that is simultaneously
+performant and maintainable while still being user-friendly and
 logically structured. Using a `Python` front-end and dancing across language
 barriers also hinders the ability to parallelize tasks down to 
-event level, the existing usage often relies on chunk or even file level
+event level, as the existing usage often relies on chunk or even file level
 parallelization. Finally, newer techniques such as automatic differentiation
-also works more smoothly without language barriers, allowing physicists to develope
+also work more smoothly without language barriers, allowing physicists to develop
 algorithms. With Julia's active auto diff community [^1], we expect 
-`UnROOT.jl` to be one of the pillar stone for physicists.
+`UnROOT.jl` to be one of the cornerstones for physicists.
 
 `UnROOT.jl` attempts to solve all of the above by choosing Julia, a
-high-performance language with simple and expressive syntax [@Julia]. Julia is
-designed to solve the two-language problem in general. This has been studied for
-HEP specifically as well [@JuliaPerformance]. Analysis software written in Julia
+high-performance language with a simple and expressive syntax [@Julia]. Julia is
+designed to solve the two-language problem in general. This has also been studied for
+HEP specifically [@JuliaPerformance]. Analysis software written in Julia
 can freely escape to a `for-loop` whenever vectorized-style processing is not
[…]
