+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Julia,
-  author = {Bezanson, Jeff. and Edelman, Alan. and Karpinski, Stefan. and Shah, Viral B.},
+  author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
   title = {Julia: A Fresh Approach to Numerical Computing},
   journal = {SIAM Review},
   volume = {59},
@@ -12,7 +12,7 @@
 }
 
 @article{JuliaPerformance,
-  title={Performance of julia for high energy physics analyses},
+  title={Performance of {J}ulia for high energy physics analyses},
   author={Stanitzki, Marcel and Strube, Jan},
   journal={Computing and Software for Big Science},
   volume={5},
@@ -109,7 +109,7 @@
 @article{Brun:1997pa,
     author = "Brun, R. and Rademakers, F.",
     editor = "Werlen, M. and Perret-Gallix, D.",
-    title = "{ROOT: An object oriented data analysis framework}",
+    title = {ROOT: An object oriented data analysis framework},
     doi = "10.1016/S0168-9002(97)00048-X",
     journal = "Nucl. Instrum. Meth. A",
     volume = "389",
