### 2022-08-10 #4452 @danielskatz
@tamasgal - I suggest a number of changes in the paper and one in the bib in https://github.com/JuliaHEP/UnROOT.jl/pull/179
