### 2022-08-18 0f6210b
Bump version for JOSS release
