### 2022-08-18 #4452 @tamasgal
Thanks @danielskatz and sorry for the late response (lot of traveling lately).

I fixed the Zenodo entry and created and archived a new version of UnROOT: 0.8.11

Here is the DOI: `10.5281/zenodo.7008076`
