### 2022-10-11 #4452 @PerilousApricot
Sorry for the delay. I was in Portugal and on vacation, flying back
yesterday. I'm back today, we able to get a julia environment installed to
verify the functionality works, so I should be able to sign off soon. Sorry
again!

On Wed, Jul 13, 2022 at 12:20 PM Daniel S. Katz ***@***.***>
wrote:

> 👋 @PerilousApricot <https://github.com/PerilousApricot> - checking in
> again...
>
